[Build, Gradle] Bundle compiler modules into KGP
Temporary limits the language version for the bundled modules to 2.0. The reasoning is well described here: https://github.com/JetBrains/kotlin/commit/9fd144563156ce08ab5d7f0c9665770458041856
Removes the runtime dependency on kotlin-compiler-embeddable and instead packs only the required modules into KGP. The next goal is to reduce the number of required modules to zero.
Also, removes an unused import for test project's buildscript as these symbols are not anymore accessible there (and never were intended to be accessible)
^KT-61706 In Progress