[AA] Drop "typeParameterSymbols.size == argumentTypes.size" contract
...assertion. In general, there is no such contract in compiler FE
(despite that in some cases Compiler FE fills missing arguments with
dummy ConeErrorTypes). Also, from the reports stack traces it seems like
it happens in invalid code. So it's decided to drop this check and
return an empty sequence, like we currently do for error types, and
cover found cases with tests.
Doing partial substitution could also be an option, but we don't know
which type argument corresponds to which type parameter.

^KT-65803