[Analysis API] KaSymbolProvider: drop `anonymousSymbol` API
This API is not correct as it throws an exception on an attempt to call
the property on regular functions.
Currently, only `KtClassOrObject` has a special name for `symbol` API
due to `KtEnumEntry` in the hierarchy.
So, it is better to drop `anonymousSymbol` completely and, potentially,
rethink it later if this dedicated API will be really needed.

^KT-70206 Fixed