[CFG] Introduce a new node and edge for variables captured by value
Lambdas capture vals and vars differently. When a var is captured, it is
captured by reference as it may change during the execution of the
lambda. However, most vals are captured by value as they do not change.
This capture by value happens when the lambda is created, not executed,
so a val must be initialized when the lambda is defined. Right now the
control-flow graph is constructed to only require the val be initialized
when the lambda is executed. Introduce a new node and edge label for
tracking variables which are captured by value when a lambda is created.

^KT-70133 Fixed